split squad example in two
